refactor: 코드 및 컴포넌트 파일 분할
